# backend/docs/swagger.yaml
basePath: /
definitions:
...
  datauptwo_app_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  datauptwo_app_dto.UserGroupCreateReq:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 256
        type: string
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
    required:
    - code
    - name
    type: object
  datauptwo_app_dto.UserGroupUpdateReq:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 256
        type: string
      id:
        type: integer
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
    required:
    - code
    - id
    - name
    type: object
...
host: localhost:18421
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: datauptwo 数据管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: datauptwo API
  version: "1.0"
paths:
...
  /api/user-groups:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 用户组列表
      tags:
      - UserGroup
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户组信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.UserGroupCreateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 创建用户组
      tags:
      - UserGroup
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户组信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.UserGroupUpdateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 更新用户组
      tags:
      - UserGroup
  /api/user-groups/{id}:
    delete:
      parameters:
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 删除用户组
      tags:
      - UserGroup
    get:
      parameters:
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 获取用户组详情
      tags:
      - UserGroup
  /api/user-groups/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 获取所有用户组
      tags:
      - UserGroup
...
swagger: "2.0"